refactor test to run with different error
